+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 Arduino JSON: change log
 ========================
 
-HEAD
+v4.4
 ----
 
+* Added `JsonArray::measureLength()` and `JsonObject::measureLength()` (issue #75)
+
+v4.3
+----
+
+* Added `JsonArray::removeAt()` to remove an element of an array (issue #58)
 * Fixed stack-overflow in `DynamicJsonBuffer` when parsing huge JSON files (issue #65)
 * Fixed wrong return value of `parseArray()` and `parseObject()` when allocation fails (issue #68)